:bug: fix: storybook을 확인할 수 없는 에러 수정
