+++ src/components/CardExpirationDate/CardExpirationDate.jsx
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { CardInfoContext } from 'App';
+import { CardInfoContext } from 'CardInfoContextProvider';
 
 import { MONTH, LIMIT_LENGTH, ACTION } from 'constants';
 import { inputNumberOnly, limitInputLength } from 'utils';
