+++ src/pages/CardAddPage.jsx
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { CardInfoContext } from 'App';
+import { CardInfoContext } from 'CardInfoContextProvider';
 
 import Header from '../components/Header/Header';
 import Card from 'components/common/Card/Card';
